Change include conditionals
When using the jinja2_native option the files lookups will
correctly return `null` when no match is found. This lead to
the playbook crashing for such cases. This commit changes
the conditions to simply check the truthyness of the result.
Which could be a greater than 0 length string when a match was found,
or either an empty string (jinja2_native=false)
or `null` (jinja2_native=true) when no match was found.